requirements-platform-service.txt
Django==5.2.9
djangorestframework==3.16.1
PyMySQL==1.0.2
Jinja2==3.1.0
MarkupSafe==3.0.3
pydantic==2.7.0
annotated-types==0.7.0
pydantic-core==2.18.1
typing-extensions==4.15.0
PyYAML==6.0.1
pytest==8.3.4
pytest-django==4.11.1
pytest-asyncio==0.25.3
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
colorama==0.4.6
iniconfig==2.0.0
packaging==25.0
pluggy==1.5.0
